+++ golang/lib/execution/executable_module_service_impl.go
@@ -22,7 +22,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis-engine-api-lib/golang/lib/networks"
 	"github.com/kurtosis-tech/kurtosis-module-api-lib/golang/kurtosis_module_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-module-api-lib/golang/lib/kurtosis_modules"
-	"github.com/palantir/stacktrace"
+	"github.com/kurtosis-tech/stacktrace"
 	"google.golang.org/protobuf/types/known/emptypb"
 )