+++ docs/changelog.md
@@ -5,7 +5,8 @@
 ### Breaking Changes
 * The [Kurtosis Client dependency](https://github.com/kurtosis-tech/kurtosis-client) has been replaced with a [Kurtosis Engine API Lib](https://github.com/kurtosis-tech/kurtosis-engine-api-lib) dependency
     * Users will need to update their import statements
-* The `ApiContainerSocketEnvVar` variable has been renamed `EngineUrlEnvVar`
+* The `KURTOSIS_API_SOCKET` envvar has been renamed to `ENGINE_SOCKET`, captured in a rename of the `ApiContainerSocketEnvVar` constant to `EngineUrlEnvVar`
+* The module now requires new `ENCLAVE_ID` envvar to start, captured in the `EnclaveIDEnvVar` constant
 
 # 0.11.1
 ### Changes